[feat] 푸드트럭 승인 상태 변경 API 를 users/admin 하위에 모아서 관리 (#32)
