+++ src/main/java/konkuk/chacall/domain/foodtruck/application/command/FoodTruckCommandService.java
@@ -2,12 +2,9 @@ package konkuk.chacall.domain.foodtruck.application.command;
 
 import konkuk.chacall.domain.foodtruck.domain.model.FoodTruck;
 import konkuk.chacall.domain.foodtruck.domain.repository.FoodTruckRepository;
-import konkuk.chacall.domain.foodtruck.presentation.dto.request.ApproveFoodTruckStatusRequest;
 import konkuk.chacall.domain.foodtruck.presentation.dto.request.FoodTruckSearchRequest;
 import konkuk.chacall.domain.foodtruck.presentation.dto.response.FoodTruckResponse;
 import konkuk.chacall.global.common.dto.CursorPagingResponse;
-import konkuk.chacall.global.common.exception.EntityNotFoundException;
-import konkuk.chacall.global.common.exception.code.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
@@ -36,13 +33,5 @@ public class FoodTruckCommandService {
 
     public boolean isNameDuplicated(String name) {
         return foodTruckRepository.existsByNameIgnoreCase(name);
-    }
-
-    // 운영자 가 사용할 기능 - 소유권 검증 필요 X
-    public void approveFoodTruckStatus(Long foodTruckId, ApproveFoodTruckStatusRequest request) {
-        FoodTruck foodTruck = foodTruckRepository.findById(foodTruckId)
-                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.FOOD_TRUCK_NOT_FOUND));
-
-        foodTruck.approveFoodTruck(request.status());
     }
 }
