+++ src/main/java/konkuk/chacall/domain/foodtruck/presentation/FoodTruckController.java
@@ -4,7 +4,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import konkuk.chacall.domain.foodtruck.application.FoodTruckService;
-import konkuk.chacall.domain.foodtruck.presentation.dto.request.ApproveFoodTruckStatusRequest;
 import konkuk.chacall.domain.foodtruck.presentation.dto.request.FoodTruckNameDuplicateCheckRequest;
 import konkuk.chacall.domain.foodtruck.presentation.dto.request.FoodTruckSearchRequest;
 import konkuk.chacall.domain.foodtruck.presentation.dto.response.FoodTruckNameDuplicateCheckResponse;
@@ -49,19 +48,5 @@ public class FoodTruckController {
             @Valid @RequestBody final FoodTruckNameDuplicateCheckRequest request
     ) {
         return BaseResponse.ok(foodTruckService.isNameDuplicated(request));
-    }
-
-    @Operation(
-            summary = "푸드트럭 승인 상태 변경",
-            description = "운영자 - 푸드트럭 승인 상태를 변경합니다."
-    )
-    @ExceptionDescription(SwaggerResponseDescription.APPROVE_FOOD_TRUCK_STATUS)
-    @PatchMapping("/{foodTruckId}/approval")
-    public BaseResponse<Void> approveFoodTruckStatus (
-            @PathVariable final Long foodTruckId,
-            @Valid @RequestBody final ApproveFoodTruckStatusRequest request
-    ) {
-        foodTruckService.approveFoodTruckStatus(foodTruckId, request);
-        return BaseResponse.ok(null);
     }
 }
